chore(backend): address comments
