+++ packages/backend/src/accounting/tigerbeetle/utils.ts
@@ -1,5 +1,10 @@
 import { validateId } from '../../shared/utils'
-import { AccountId, uuidToBigInt } from '../utils'
+
+export type AccountId = string | number | bigint
+
+export function uuidToBigInt(id: string): bigint {
+  return BigInt(`0x${id.replace(/-/g, '')}`)
+}
 
 export function toTigerbeetleId(id: AccountId): bigint {
   if (typeof id === 'number') {
