--- packages/backend/src/accounting/utils.ts
@@ -1,5 +1,0 @@
-export type AccountId = string | number | bigint
-
-export function uuidToBigInt(id: string): bigint {
-  return BigInt(`0x${id.replace(/-/g, '')}`)
-}
